VPDataAPI/requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.6
djangorestframework==3.12.4
drf-spectacular==0.17.3
inflection==0.5.1
jsonschema==3.2.0
numpy==1.21.1
pandas==1.3.2
pandas-stubs==1.2.0.1
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
